+++ .env_example
@@ -1,8 +1,8 @@
-STEAM_API_KEY=REPLACE_THIS,REPLACE_THIS
+STEAM_API_KEY=REPLACE_THIS
 RECAPTCHA_PUBLIC_KEY=REPLACE_THIS
 RECAPTCHA_SECRET_KEY=REPLACE_THIS
-STEAM_USER=REPLACE_THIS,REPLACE_THIS
-STEAM_PASS=REPLACE_THIS,REPLACE_THIS
+STEAM_USER=REPLACE_THIS
+STEAM_PASS=REPLACE_THIS
 
 SESSION_SECRET=somesecretvalue
 ROOT_URL=http://localhost:5000
