+++ .gitignore
@@ -1,19 +1,15 @@
 .env
 node_modules
-replays
-replays_test
+replays*
 *.log
 sentry
 servers
 *.dem
 *.bz2
 output.json
-uploads
 coverage
-bundle.js
-.bundle.js
-constants.json
-constants_test.json
+*bundle.js
+constants*.json
 parser/target
 !test/*
 /bin/
