+++ parser.js
@@ -2,4 +2,4 @@ var utility = require('./utility');
 var processors = require('./processors');
 var jobs = utility.jobs;
 console.log("[PARSER] starting parser");
-jobs.process('parse', process.env.STREAM ? processors.processParseStream : processors.processParse);
+jobs.process('parse', processors.processParse);
