+++ runner.js
@@ -1,6 +1,6 @@
 var dotenv = require('dotenv');
 dotenv.load();
 var args = process.argv.slice(2);
-require('./' + args[0])(function(err, res) {
+require('./tasks/' + args[0])(function(err, res) {
     process.exit(err);
 });
