+++ selector.js
@@ -1,0 +1,22 @@
+var moment = require('moment');
+
+module.exports = function(type) {
+    var opts = {
+        "untrack": {
+            track: 1,
+            last_visited: {
+                $lt: moment().subtract(3, 'day').toDate()
+            }
+        },
+        "fullhistory": {
+            track: 1,
+            join_date: {
+                $lt: moment().subtract(7, 'day').toDate()
+            },
+            last_visited: {
+                $gt: moment().subtract(1, 'day').toDate()
+            }
+        }
+    };
+    return opts[type];
+};
