+++ tasks/updatenames.js
@@ -1,8 +1,8 @@
-var utility = require('./utility');
-var db = utility.db;
+var utility = require('../utility');
+var db = require('../db');
 var async = require('async');
 var queueReq = utility.queueReq;
-var constants = require('./sources.json');
+var constants = require('../sources.json');
 module.exports = function updateNames(cb) {
     var buckets = 1; //do only some of the names at once
     var target = Math.floor(Math.random() * buckets);
