+++ views/about.jade
@@ -31,7 +31,7 @@ block content
   p YASP only starts tracking your matches from when you signed in.  We get full match histories for contributors and active YASP users.
   
   h3 How do untracking and full history eligibility work?
-  p Currently, we untrack users who have not visited YASP for 3 days.  For full history, we select one user every hour whose account is at least 10 days old, is tracked, and has visited in the last day.
+  p Currently, we untrack users who have not visited YASP for 3 days.  For full history, we select one user every hour whose account is at least 7 days old, is tracked, and has visited in the last day.
   
   h3 Why am I still missing some matches after getting my full history?
   p The Valve API limits results to a maximum of 500.  We can get 500 games per hero per player, but if you have more than 500 games on a hero, we won't be able to get the rest.  Sorry.
